backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# NLP and ML
transformers==4.36.0
torch>=2.0.0
spacy==3.7.2
nltk==3.8.1
scikit-learn==1.4.0

# OCR
easyocr==1.7.1
pytesseract==0.3.10
Pillow>=10.0.0

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
openpyxl==3.1.2
pandas==2.1.4
python-docx==1.1.0

# Database
motor==3.3.2
pymongo==4.6.1
redis==5.0.1

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
aiofiles==23.2.1

# Graph and Analysis
networkx==3.2.1
matplotlib==3.8.2
scipy>=1.11.0
numpy>=1.24.0

# Testing
pytest==7.4.3
httpx==0.26.0